+++ src/request.js
@@ -60,6 +60,17 @@ export default class Request {
     }
 
     /**
+     * @param {object} config
+     *
+     * @returns {this} The current instance.
+     */
+    setConfig (config) {
+        this._connection.setConfig(_.merge(this._connection.getConfig(), config));
+
+        return this;
+    }
+
+    /**
      * @returns {this} The current instance.
      */
     cancel () {
@@ -154,7 +165,7 @@ export default class Request {
 
         this._connection.cancelRunning(this);
 
-        this._connection.setConfig({
+        this.setConfig({
             headers: {
                 'Content-Type': 'multipart/form-data',
             },